Fix tag broken tests and create test helper
